refactoring - split networkGraph into separate component
